Fixed the Obj Files

+++ RasterizerDemo/SceneD3D11.h
@@ -34,7 +34,7 @@ enum class RenderMode
 };
 
 struct RenderType {
-    float renderType;
+    int renderType;
     DirectX::XMFLOAT3 padding;
 };
